LayoutTests/imported/w3c/web-platform-tests/css/css-tables/table-model-fixup-2-expected.txt

Table fixup

Check that replaced elements display types are handled properly in fixup

Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top)


Both text inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- img elements


Both text inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=text elements


Both button inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=button elements


Both file inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=file elements


Both images should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- img elements



PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top) assert_equals: expected 808.875 but got 840.875
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top) assert_equals: expected 1561.875 but got 1593.875
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top) assert_equals: expected 2304.875 but got 2336.875
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=text elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=button elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=file elements
PASS Replaced elements outside a table cannot be table and are considered block -- img elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements assert_equals: expected 3748.21875 but got 3766.21875
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- img elements assert_equals: expected 3880.21875 but got 3912.21875
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements assert_equals: expected 4272.21875 but got 4290.21875
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements assert_equals: expected 4404.21875 but got 4436.21875
PASS Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements assert_equals: expected 4796.21875 but got 4814.21875
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- img elements assert_equals: expected 4928.21875 but got 4960.21875
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements assert_equals: expected 5320.21875 but got 5338.21875
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- img elements assert_equals: expected 5452.21875 but got 5484.21875
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements assert_equals: expected 5844.21875 but got 5862.21875
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- img elements assert_equals: expected 5976.21875 but got 6008.21875
